third_party/blink/web_tests/platform/win/external/wpt/html/syntax/speculative-parsing/generated/page-load/svg-script-src.tentative-expected.txt

This is a testharness.js-based test.
FAIL Speculative parsing, page load: svg-script-src Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nHost: web-platform.test:8001\nConnection: keep-alive\nUser-Agent: Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/999.77.34.5 Safari/537.36\nAccept: */*\nReferer: http://web-platform.test:8001/html/syntax/speculative-parsing/generated/page-load/resources/svg-script-src-framed.sub.html?uuid=8a1a7a36-c301-4d4a-baf7-9ac811770a1a\nAccept-Encoding: gzip, deflate\nAccept-Language: en-us,en\n\n"
Harness: the test ran to completion.
